refactor: extract cursor state management into useCursor hook
Consolidate cursor/selectedIndex logic scattered across useClipboardSearch
and App.tsx into a dedicated useCursor hook with reset, clamp, moveNext,
movePrev, and selectByIndex methods.
